discord_bot/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
discord.py==2.5.2
frozenlist==1.7.0
idna==3.10
markdownify==0.13.1
multidict==6.6.3
numpy==1.26.3
pandas==2.3.1
propcache==0.3.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jobspy==1.1.80
pytz==2025.2
regex==2024.11.6
requests==2.32.4
six==1.17.0
soupsieve==2.7
tls-client==1.0.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uv==0.7.20
yarl==1.20.1
